mac80211 injection: sendmsg() with MSG_DONTWAIT
Without a thorough research, it seems that under sendmsg() there is a wireless driver queue for
injected frames, of unknown length. It seems that whenever sendmsg() is called, the actual work it
does, apart of some frame verification, is appending the frame to this queue. However, if the queue
is full, the sendmsg() call would block until there is some space.

In case there is a big amount of background traffic in the network neighborhood, the packet
injection queue would move very slow. Thus, the sendmsg() could block even for low amounts of
locally generated traffic, and such situation is independent of the activity of local system.

Blocking sendmsg() makes a disaster in the scheduler, so turn on the MSG_DONTWAIT option.